requirements.txt
Django==5.0.6
django-bootstrap5==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
djoser==2.3.1
django-simple-history==3.8.0
django-filter==21.1
psycopg[binary]==3.2.6
uvicorn==0.34.0
gunicorn==23.0.0
requests==2.32.3
channels==4.1.0
channels-redis==4.3.0
websockets==15.0.1